backend/requirements.txt
# Core Automation
selenium>=4.15.0
webdriver-manager>=4.0.0
pyautogui>=0.9.54

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Machine Learning / NLP
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Database
aiosqlite>=0.19.0

# UI Dashboard
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Text-to-Speech (optional)
pyttsx3>=2.90

# Progress Tracking
tqdm>=4.66.0
